requirements.txt
# The Virtual Brain and dependencies
tvb-library>=2.8
tvb-data>=2.0
tvb-gdist>=2.1.0

# Scientific computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Neuroimaging
nibabel>=3.2.0
nilearn>=0.9.0

# Utilities
h5py>=3.1.0
jupyter>=1.0.0
ipywidgets>=7.6.0
tqdm>=4.62.0
